refactor(distributor): extract profile stripping into sampling package
Move stripProfileToTotals out of distributor.go into a Stripper component
in pkg/distributor/sampling, called as d.stripper.StripToTotals. Upcoming
work will make stripping recording-rules aware, which means more stripping
mechanism; this gives it a home of its own so distributor.go stops growing.

The strip-semantics unit tests move next to the code in the sampling
package. The distributor tests keep verifying the accounting around
stripping (discarded bytes, sampled label, language detection ordering)
against an explicit golden fixture.
