+++ pkg/distributor/distributor.go
@@ -114,6 +114,7 @@ type Distributor struct {
 	asyncRequests          sync.WaitGroup
 	ingestionLimitsSampler *ingestlimits.Sampler
 	usageGroupEvaluator    *validation.UsageGroupEvaluator
+	stripper               *sampling.Stripper
 
 	subservices        *services.Manager
 	subservicesWatcher *services.FailureWatcher
@@ -195,6 +196,7 @@ func New(
 		profileReceivedStats:    usagestats.NewMultiCounter("distributor_profiles_received", "lang"),
 		profileScopeStats:       usagestats.NewMultiCounter("distributor_profiles_received_by_scope", "scope"),
 		profileSizeStats:        usagestats.NewMultiStatistics("distributor_profile_sizes", "lang"),
+		stripper:                sampling.NewStripper(),
 	}
 
 	ingesterRoute := writepath.IngesterFunc(d.sendRequestsToIngester)
@@ -609,7 +611,7 @@ func (d *Distributor) pushSeries(ctx context.Context, req *distributormodel.Prof
 		// Language detection reads the string table, which is about to be
 		// stripped; the result is cached in the request.
 		d.GetProfileLanguage(req)
-		stripProfileToTotals(req.Profile.Profile)
+		d.stripper.StripToTotals(req.Profile.Profile)
 		req.Labels = phlaremodel.Labels(req.Labels).InsertSorted(phlaremodel.LabelNameSampled, "true")
 
 		// The stripped part of the profile is discarded, and from here on
@@ -951,70 +953,6 @@ func (d *Distributor) sendRequestsToSegmentWriter(ctx context.Context, req *dist
 	}
 
 	return connect.NewResponse(&pushv1.PushResponse{}), nil
-}
-
-// stripProfileToTotals reduces the profile to a single sample holding the
-// sum of all sample values: stacktraces, symbols, and sample labels are
-// dropped, only the series totals are kept.
-func stripProfileToTotals(p *profilev1.Profile) {
-	var total *profilev1.Sample
-	for _, s := range p.Sample {
-		// Samples that Normalize would drop (value length mismatch,
-		// negative values) must not contribute to the totals.
-		if len(s.Value) != len(p.SampleType) || hasNegativeValue(s) {
-			continue
-		}
-		if total == nil {
-			total = &profilev1.Sample{Value: make([]int64, len(p.SampleType))}
-		}
-		for i, v := range s.Value {
-			total.Value[i] += v
-		}
-	}
-	p.Sample = nil
-	if total != nil {
-		p.Sample = []*profilev1.Sample{total}
-	}
-	p.Location = nil
-	p.Function = nil
-	p.Mapping = nil
-
-	oldStrings := p.StringTable
-	newStrings := []string{""}
-	remap := map[int64]int64{0: 0}
-	intern := func(old int64) int64 {
-		if n, ok := remap[old]; ok {
-			return n
-		}
-		n := int64(len(newStrings))
-		newStrings = append(newStrings, oldStrings[old])
-		remap[old] = n
-		return n
-	}
-	p.DropFrames = intern(p.DropFrames)
-	p.KeepFrames = intern(p.KeepFrames)
-	p.DefaultSampleType = intern(p.DefaultSampleType)
-	for _, vt := range p.SampleType {
-		vt.Type = intern(vt.Type)
-		vt.Unit = intern(vt.Unit)
-	}
-	if p.PeriodType != nil {
-		p.PeriodType.Type = intern(p.PeriodType.Type)
-		p.PeriodType.Unit = intern(p.PeriodType.Unit)
-	}
-	for i, c := range p.Comment {
-		p.Comment[i] = intern(c)
-	}
-	p.StringTable = newStrings
-}
-
-func hasNegativeValue(s *profilev1.Sample) bool {
-	for _, v := range s.Value {
-		if v < 0 {
-			return true
-		}
-	}
-	return false
 }
 
 // profileSizeBytes returns the size of symbols and samples in bytes from a given fullSize.
